Move paste logic to schemas
